refs(requests): Move `validate_request_content` into the `request` module
This allows `validate_request_content` to be used in subscriptions without circular import issues.
